+++ src/stories/shared-mock-repositories/info/DataverseInfoMockRepository.ts
@@ -1,12 +1,15 @@
 import { FileSizeUnit } from '@/files/domain/models/FileMetadata'
+import { DatasetMetadataExportFormats } from '@/info/domain/models/DatasetMetadataExportFormats'
 import { DataverseVersion } from '@/info/domain/models/DataverseVersion'
 import { TermsOfUse } from '@/info/domain/models/TermsOfUse'
 import { DataverseInfoRepository } from '@/info/domain/repositories/DataverseInfoRepository'
 import { Setting } from '@/settings/domain/models/Setting'
 import { ZipDownloadLimit } from '@/settings/domain/models/ZipDownloadLimit'
-import { DataverseVersionMother } from '@tests/component/info/models/DataverseVersionMother'
-import { TermsOfUseMother } from '@tests/component/info/models/TermsOfUseMother'
+import { DatasetMetadataExportFormatsMother } from '@tests/component/info/domain/models/DatasetMetadataExportFormatsMother'
+import { DataverseVersionMother } from '@tests/component/info/domain/models/DataverseVersionMother'
+import { TermsOfUseMother } from '@tests/component/info/domain/models/TermsOfUseMother'
 import { SettingMother } from '@tests/component/settings/domain/models/SettingMother'
+import { FakerHelper } from '@tests/component/shared/FakerHelper'
 
 export class DataverseInfoMockRepository implements DataverseInfoRepository {
   getVersion(): Promise<DataverseVersion> {
@@ -62,6 +65,14 @@ export class DataverseInfoMockRepository implements DataverseInfoRepository {
           ])
         )
       }, 1_000)
+    })
+  }
+
+  getAvailableDatasetMetadataExportFormats(): Promise<DatasetMetadataExportFormats> {
+    return new Promise((resolve) => {
+      setTimeout(() => {
+        resolve(DatasetMetadataExportFormatsMother.create())
+      }, FakerHelper.loadingTimout())
     })
   }
 }