+++ tests/component/sections/dataset/dataset-action-buttons/access-dataset-menu/AccessDatasetMenu.spec.tsx
@@ -8,6 +8,7 @@ import {
 import { FileSizeUnit } from '../../../../../../src/files/domain/models/FileMetadata'
 
 const downloadUrls = DatasetDownloadUrlsMother.create()
+
 describe('AccessDatasetMenu', () => {
   it('renders the AccessDatasetMenu if the user has download files permissions and the dataset is not deaccessioned', () => {
     const version = DatasetVersionMother.createReleased()
@@ -24,6 +25,7 @@ describe('AccessDatasetMenu', () => {
         permissions={permissions}
         downloadUrls={downloadUrls}
         fileStore="s3"
+        persistentId="doi:10.5072/FK2/ABCDEFGH"
       />
     )
 
@@ -48,6 +50,7 @@ describe('AccessDatasetMenu', () => {
         permissions={permissions}
         downloadUrls={downloadUrls}
         fileStore="s3"
+        persistentId="doi:10.5072/FK2/ABCDEFGH"
       />
     )
     cy.findByRole('button', { name: 'Access Dataset' }).should('exist')
@@ -68,6 +71,7 @@ describe('AccessDatasetMenu', () => {
         permissions={permissions}
         downloadUrls={downloadUrls}
         fileStore="s3"
+        persistentId="doi:10.5072/FK2/ABCDEFGH"
       />
     )
     cy.findByRole('button', { name: 'Access Dataset' }).should('not.exist')
@@ -88,6 +92,7 @@ describe('AccessDatasetMenu', () => {
         permissions={permissions}
         downloadUrls={downloadUrls}
         fileStore="s3"
+        persistentId="doi:10.5072/FK2/ABCDEFGH"
       />
     )
     cy.findByRole('button', { name: 'Access Dataset' }).should('not.exist')
@@ -107,6 +112,7 @@ describe('AccessDatasetMenu', () => {
         permissions={permissions}
         downloadUrls={downloadUrls}
         fileStore="s3"
+        persistentId="doi:10.5072/FK2/ABCDEFGH"
       />
     )
     cy.findByRole('button', { name: 'Access Dataset' }).should('exist')
@@ -134,6 +140,7 @@ describe('AccessDatasetMenu', () => {
         permissions={permissions}
         downloadUrls={downloadUrls}
         fileStore="s3"
+        persistentId="doi:10.5072/FK2/ABCDEFGH"
       />
     )
     cy.findByRole('button', { name: 'Access Dataset' }).should('exist')
@@ -164,6 +171,7 @@ describe('AccessDatasetMenu', () => {
         permissions={permissions}
         downloadUrls={downloadUrls}
         fileStore="s3"
+        persistentId="doi:10.5072/FK2/ABCDEFGH"
       />
     )
     cy.findByRole('button', { name: 'Access Dataset' }).should('not.exist')
@@ -184,6 +192,7 @@ describe('AccessDatasetMenu', () => {
         permissions={permissions}
         downloadUrls={downloadUrls}
         fileStore="not-s3"
+        persistentId="doi:10.5072/FK2/ABCDEFGH"
       />
     )
     cy.findByRole('button', { name: 'Access Dataset' }).should('not.exist')